Refactor command initialization
Now it's possible to load plugins from sources configured in .myks.yaml
