+++ cmd/root.go
@@ -1,14 +1,12 @@
 package cmd
 
 import (
-	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
 	"strings"
 	"time"
 
-	aurora "github.com/logrusorgru/aurora/v4"
 	"github.com/rs/zerolog"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/cobra"
@@ -18,10 +16,8 @@ import (
 )
 
 const (
-	ANNOTATION_SMART_MODE = "feat:smart-mode"
-	ANNOTATION_TRUE       = "true"
-	MYKS_CONFIG_NAME      = ".myks"
-	MYKS_CONFIG_TYPE      = "yaml"
+	MYKS_CONFIG_NAME = ".myks"
+	MYKS_CONFIG_TYPE = "yaml"
 )
 
 var (
@@ -30,24 +26,38 @@ var (
 	asyncLevel int
 )
 
-var rootCmd = &cobra.Command{
-	Use:   "myks",
-	Short: "Myks helps to manage configuration for kubernetes clusters",
-	Long: `Myks fetches K8s workloads from a variety of sources, e.g. Helm charts or Git Repositories. It renders their respective yaml files to the file system in a structure of environments and their applications.
-
-It supports prototype applications that can be shared between environments and inheritance of configuration from parent environments to their "children".
-
-Myks supports two positional arguments:
-
-- A comma-separated list of environments to render. If you provide "ALL", all environments will be rendered.
-- A comma-separated list of applications to render. If you don't provide this argument or provide "ALL", all applications will be rendered.
-
-If you do not provide any positional arguments, myks will run in "Smart Mode". In Smart Mode, myks will only render environments and applications that have changed since the last run.
-`,
+func NewMyksCmd(version, commit, date string) *cobra.Command {
+	cmd := newRootCmd(version, commit, date)
+	cmd.AddCommand(allCmd)
+	cmd.AddCommand(renderCmd)
+	cmd.AddCommand(newInitCmd())
+	cmd.AddCommand(newPrintConfigCmd())
+	cmd.AddCommand(newSyncCmd())
+	initConfig()
+	addPlugins(cmd)
+	return cmd
 }
 
-func init() {
-	cobra.OnInitialize(initConfig)
+func newRootCmd(version, commit, date string) *cobra.Command {
+	var rootCmd = &cobra.Command{
+		Use:   "myks",
+		Short: "Myks helps to manage configuration for kubernetes clusters",
+		Long: `Myks fetches K8s workloads from a variety of sources, e.g. Helm charts or Git Repositories. It renders their respective yaml files to the file system in a structure of environments and their applications.
+	
+	It supports prototype applications that can be shared between environments and inheritance of configuration from parent environments to their "children".
+	
+	Myks supports two positional arguments:
+	
+	- A comma-separated list of environments to render. If you provide "ALL", all environments will be rendered.
+	- A comma-separated list of applications to render. If you don't provide this argument or provide "ALL", all applications will be rendered.
+	
+	If you do not provide any positional arguments, myks will run in "Smart Mode". In Smart Mode, myks will only render environments and applications that have changed since the last run.
+	`,
+	}
+
+	rootCmd.Version = fmt.Sprintf(`%s
+     commit: %s
+     date:   %s`, version, commit, date)
 
 	rootCmd.PersistentFlags().StringP("log-level", "l", "info", "set the logging level")
 
@@ -77,74 +87,7 @@ func init() {
 		}
 		return initTargetEnvsAndApps(cmd, args)
 	}
-}
-
-func initTargetEnvsAndApps(cmd *cobra.Command, args []string) (err error) {
-	// Check positional arguments for Smart Mode:
-	// 1. Comma-separated list of environment search paths or ALL to search everywhere (default: ALL)
-	// 2. Comma-separated list of application names or none to process all applications (default: none)
-
-	switch len(args) {
-	case 0:
-		// smart mode requires instantiation of globe object to get the list of environments
-		// the globe object will not be used later in the process. It is only used to get the list of all environments and their apps.
-		globeAllEnvsAndApps := myks.New(".")
-		envAppMap, err = globeAllEnvsAndApps.DetectChangedEnvsAndApps(viper.GetString("smart-mode.base-revision"))
-		if err != nil {
-			log.Warn().Err(err).Msg("Unable to run Smart Mode. Rendering everything.")
-		} else if len(envAppMap) == 0 {
-			log.Warn().Msg("Smart Mode did not find any changes. Exiting.")
-			os.Exit(0)
-		}
-	case 1:
-		if args[0] != "ALL" {
-			envAppMap = make(myks.EnvAppMap)
-			for _, env := range strings.Split(args[0], ",") {
-				envAppMap[env] = nil
-			}
-		}
-	case 2:
-		var appNames []string
-		if args[1] != "ALL" {
-			appNames = strings.Split(args[1], ",")
-		}
-
-		envAppMap = make(myks.EnvAppMap)
-		if args[0] != "ALL" {
-			for _, env := range strings.Split(args[0], ",") {
-				envAppMap[env] = appNames
-			}
-		} else {
-			// TODO: Use Globe.EnvironmentBaseDir instead of the hardcoded key
-			envAppMap["envs"] = appNames
-		}
-
-	default:
-		err := errors.New("too many positional arguments")
-		log.Error().Err(err).Msg("Unable to parse positional arguments")
-		return err
-	}
-
-	log.Debug().
-		Interface("envAppMap", envAppMap).
-		Msg("Parsed arguments")
-
-	if viper.GetBool("smart-mode.only-print") {
-		fmt.Println(aurora.Bold("\nSmart Mode detected:"))
-		for env, apps := range envAppMap {
-			fmt.Printf("→ %s\n", env)
-			if apps == nil {
-				fmt.Println(aurora.Bold("    ALL"))
-			} else {
-				for _, app := range apps {
-					fmt.Printf("    %s\n", app)
-				}
-			}
-		}
-		os.Exit(0)
-	}
-
-	return nil
+	return rootCmd
 }
 
 func initConfig() {
@@ -192,14 +135,4 @@ func initLogger() {
 
 	log.Info().Msgf("Setting log level to: %s", logLevel)
 	zerolog.SetGlobalLevel(logLevel)
-}
-
-func SetVersionInfo(version, commit, date string) {
-	rootCmd.Version = fmt.Sprintf(`%s
-     commit: %s
-     date:   %s`, version, commit, date)
-}
-
-func Execute() {
-	cobra.CheckErr(rootCmd.Execute())
 }
