+++ internal/myks/assets/myks_config.yaml
@@ -5,3 +5,6 @@ async: 0
 # One of the zerolog log levels.
 # See: https://github.com/rs/zerolog#leveled-logging
 log-level: info
+# Load all binaries from the following folders as myks plugins.
+plugin-sources:
+    - ./plugins
