+++ main.go
@@ -17,6 +17,9 @@ var (
 
 func main() {
 	log.Logger = log.Output(zerolog.ConsoleWriter{Out: os.Stderr})
-	cmd.SetVersionInfo(version, commit, date)
-	cmd.Execute()
+	cmd := cmd.NewMyksCmd(version, commit, date)
+	err := cmd.Execute()
+	if err != nil {
+		log.Fatal().Err(err).Msg("Error executing myks")
+	}
 }
